+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.11.2 (Unreleased)
-ENHANCEMENTS:
-* provider: provider uses permanent client-request-id identifier while the terraform is running
+BUG FIXES:
+* provider: fix provider name and version detection   
 
 ## 0.11.1 (September 13, 2019)
 ENHANCEMENTS: